test(report): split 966-line test_data_integrity.py by invariant (#113)
`services/report/tests/integration/test_data_integrity.py` had 4
distinct top-level test functions, each verifying a different POS
accounting invariant. Split one file per invariant — each remains
fully self-contained (no shared module-level helpers).

- test_payment_sum_invariant.py        (126 lines)
- test_multi_transaction_integrity.py  (193 lines)
- test_store_daily_integrity.py        (275 lines)
- test_cartesian_safety.py             (347 lines)

Each new file inherits the same imports + the `set_env_vars` fixture
provided by the integration-tier conftest. No test logic changed.

Pass count unchanged: report integration tier 47/47 (was 47/47).

Refs #113
